Move generateRarity to its own file and made it accept a contract address
