+++ src/alchemy/index.js
@@ -1,5 +1,5 @@
 import { Network, Alchemy } from 'alchemy-sdk';
-import { ALCHEMY_API_KEY, CONTRACT_ADDRESS } from '../utils/constants.js';
+import { ALCHEMY_API_KEY } from '../utils/constants.js';
 
 const settings = {
   apiKey: ALCHEMY_API_KEY,
@@ -8,22 +8,22 @@ const settings = {
 
 const alchemyClient = new Alchemy(settings);
 
-const getNFTsForCollectionOnce = async (pageKey) => {
-  const response = await alchemyClient.nft.getNftsForContract(CONTRACT_ADDRESS, {
-    pageKey: pageKey,
+const getNFTsForCollectionOnce = async (contractAddress, pageKey) => {
+  const response = await alchemyClient.nft.getNftsForContract(contractAddress, {
+    pageKey,
     withMetadata: true
   });
 
   return response;
 };
 
-export const getNftsAndMetaData = async () => {
+export const getNftsAndMetaData = async (contractAddress) => {
   const metadata = [];
   const allNfts = [];
 
   let nextPage = '';
   while (nextPage || nextPage === '') {
-    const { nfts, pageKey } = await getNFTsForCollectionOnce(nextPage);
+    const { nfts, pageKey } = await getNFTsForCollectionOnce(contractAddress, nextPage);
 
     for (const token of nfts) {
       if (token.rawMetadata.attributes) {
